refactor(query): extract plan profile
