+++ src/query/catalog/src/table_context.rs
@@ -35,8 +35,8 @@ use databend_common_meta_app::principal::OnErrorMode;
 use databend_common_meta_app::principal::RoleInfo;
 use databend_common_meta_app::principal::UserDefinedConnection;
 use databend_common_meta_app::principal::UserInfo;
-use databend_common_pipeline_core::processors::profile::PlanProfile;
-use databend_common_pipeline_core::processors::profile::Profile;
+use databend_common_pipeline_core::processors::PlanProfile;
+use databend_common_pipeline_core::processors::Profile;
 use databend_common_pipeline_core::InputError;
 use databend_common_settings::Settings;
 use databend_common_storage::CopyStatus;
