+++ src/query/service/src/api/rpc/exchange/statistics_sender.rs
@@ -24,8 +24,8 @@ use databend_common_base::runtime::TrySpawn;
 use databend_common_catalog::table_context::TableContext;
 use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
-use databend_common_pipeline_core::processors::profile::PlanProfile;
 use databend_common_pipeline_core::processors::profile::Profile;
+use databend_common_pipeline_core::processors::PlanProfile;
 use databend_common_storage::MergeStatus;
 use futures_util::future::Either;
 use log::warn;
