+++ src/query/service/src/api/rpc/packets/packet_data.rs
@@ -23,7 +23,7 @@ use databend_common_arrow::arrow_format::flight::data::FlightData;
 use databend_common_catalog::statistics::data_cache_statistics::DataCacheMetricValues;
 use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
-use databend_common_pipeline_core::processors::profile::PlanProfile;
+use databend_common_pipeline_core::processors::PlanProfile;
 use databend_common_storage::CopyStatus;
 use databend_common_storage::MergeStatus;
 use log::error;
