+++ src/query/service/src/interpreters/interpreter.rs
@@ -21,7 +21,7 @@ use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
 use databend_common_expression::SendableDataBlockStream;
 use databend_common_pipeline_core::get_statistics_desc;
-use databend_common_pipeline_core::processors::profile::PlanProfile;
+use databend_common_pipeline_core::processors::PlanProfile;
 use databend_common_pipeline_core::processors::ProfileDesc;
 use databend_common_pipeline_core::processors::ProfileStatisticsName;
 use databend_common_pipeline_core::SourcePipeBuilder;
