+++ src/query/service/src/pipelines/executor/executor_graph.rs
@@ -23,9 +23,9 @@ use databend_common_base::runtime::TrySpawn;
 use databend_common_exception::ErrorCode;
 use databend_common_exception::Result;
 use databend_common_pipeline_core::processors::EventCause;
+use databend_common_pipeline_core::processors::PlanScope;
 use databend_common_pipeline_core::processors::Profile;
 use databend_common_pipeline_core::Pipeline;
-use databend_common_pipeline_core::PlanScope;
 use log::debug;
 use log::trace;
 use minitrace::prelude::*;
