+++ src/query/service/src/pipelines/pipeline_builder.rs
@@ -20,9 +20,9 @@ use databend_common_exception::Result;
 use databend_common_expression::DataField;
 use databend_common_expression::FunctionContext;
 use databend_common_pipeline_core::processors::profile::ProfileLabel;
+use databend_common_pipeline_core::processors::PlanScope;
+use databend_common_pipeline_core::processors::PlanScopeGuard;
 use databend_common_pipeline_core::Pipeline;
-use databend_common_pipeline_core::PlanScope;
-use databend_common_pipeline_core::PlanScopeGuard;
 use databend_common_settings::Settings;
 use databend_common_sql::executor::PhysicalPlan;
 use databend_common_sql::IndexType;
