+++ src/query/service/src/sessions/query_ctx.rs
@@ -69,8 +69,8 @@ use databend_common_meta_app::schema::CatalogInfo;
 use databend_common_meta_app::schema::GetTableCopiedFileReq;
 use databend_common_meta_app::schema::TableInfo;
 use databend_common_metrics::storage::*;
-use databend_common_pipeline_core::processors::profile::PlanProfile;
 use databend_common_pipeline_core::processors::profile::Profile;
+use databend_common_pipeline_core::processors::PlanProfile;
 use databend_common_pipeline_core::processors::ProfileStatisticsName;
 use databend_common_pipeline_core::InputError;
 use databend_common_settings::Settings;
